Overhaul of the queued zoning system

+++ src/tools/quest_parser/main.cpp
@@ -595,7 +595,7 @@ int main( int argc, char** argv )
     {
 
       std::string entry( &section[ offset ] );
-      offset += entry.size() + 1;
+      offset += static_cast< uint32_t >( entry.size() + 1 );
 
       if( entry.size() > 3
           && entry.find_first_not_of( "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ01234567890_-" ) ==
